requirements_min.txt
numpy==1.24.4
opencv_python==4.10.0.84
Pillow==11.1.0
torch==2.5.1
torchvision==0.20.1
pandas==2.2.3
requests==2.32.3
